feat: Add menu items configuration file
The code changes introduce a new configuration file `menu_items.dart` that defines a `MenuItem` class and a list of `appMenuItems`. This file is responsible for storing the menu items data used in the application.

The commit message suggests adding the file to the repository and includes a brief description of the purpose of the changes.