feat: Add AppTheme class for managing the application theme
The code changes introduce a new file `app_theme.dart` that defines the `AppTheme` class. This class is responsible for managing the application theme and provides a `getTheme()` method to retrieve the current theme data.

This commit adds the `app_theme.dart` file to the repository, which will be used to configure the theme of the application.