feat: Add HomeScreen widget to display the home screen
The code changes introduce a new file `home_screen.dart` that defines the `HomeScreen` widget. This widget is responsible for displaying the home screen of the application.

This commit adds the `home_screen.dart` file to the repository, which will be used to render the home screen in the application.