feat: Add navigation to ButtonsScreen from HomeScreen
The code changes in `home_screen.dart` add a navigation functionality to the `onTap` event of the `_CustomListTile` widget. When a menu item is tapped, it will navigate to the `ButtonsScreen` using `Navigator.of(context).push(MaterialPageRoute(builder: (context) => const ButtonsScreen()))`.

This commit enhances the user experience by allowing seamless navigation from the home screen to the buttons screen.