feat: Add SideMenu widget for navigation drawer functionality
This commit adds the SideMenu widget, which provides navigation drawer functionality to the application. The SideMenu widget is used in the HomeScreen to display a menu that can be accessed by swiping from the left side of the screen or tapping on a menu icon. The menu includes various options for navigating within the application.

These additions enhance the user experience by providing a convenient and intuitive way for users to access different sections of the application.